consensus: fix liveness wedge + deterministic microblock fork-choice
Wedge root: production targeted the gossip network-height, not the local applied
tip -> elected producer self-excluded on a prev it never had -> failover round
ran away -> halt with no fork.

- node.rs: produce from applied tip; bound failover round (cap 50 -> recovery)
- unified_p2p.rs: sync escape hatch -- gap recovery not gated by peer-height
  staleness (safety = block verification, not gossip freshness)
- block_pipeline.rs: a higher 2f+1-certified-round microblock supersedes a
  same-height block via the existing finality-guarded reorg
- consensus_v2_node.rs: always-on checkpoint enforce (never finalize a fork)
- block.rs: reputation as integer centipercent

Requires fresh genesis.

+++ core/qnet-state/src/block.rs
@@ -337,8 +337,9 @@ pub struct ConsensusData {
 pub struct EligibleProducer {
     /// Node identifier (e.g., "genesis_node_001" or "node_abc123")
     pub node_id: String,
-    /// Reputation score at snapshot time (0.0 - 1.0)
-    pub reputation: f64,
+    /// Reputation at snapshot time as fixed-point centipercent (u32): 70.00% = 7000, max = 10000.
+    /// Integer (not f64) so this consensus-committed value is bit-identical on every node.
+    pub reputation: u32,
 }
 
 /// Slashing event data for blockchain storage (simplified for serialization)
